Fix useMount() to avoid eslint warning
After recent updates, we're getting the `react-hooks/exhaustive-deps`
warning for `useMount()`.

Additionally, React 18 (when we upgrade to it) will call the
`useEffect()` twice in strict mode:

https://stackoverflow.com/a/75126229/10310114

So we can't rely on the "empty dependency array" approach.

To fix, include the callback function in the dependency array (to avoid
the eslint warning), but explicitly only call it once from within
`useMount()`.

Also, add deps for future migration to TypeScript.

#31